+++ src/Logary.Tests/Engine.fs
@@ -8,7 +8,7 @@ open Hopac
 open Hopac.Infixes
 open Hopac.Extensions
 open Logary.Internals
-open Logary.EventProcessing
+open Logary.Configuration
 open NodaTime
 open System.Net.NetworkInformation