+++ src/legacy/Logary.Services.Rutta/Shipper.fs
@@ -15,7 +15,6 @@ module Shipper =
   open Logary.Configuration
   open fszmq
   open fszmq.Socket
-  open Logary.EventProcessing
 
   let private runLogary shipperConf: IDisposable =
     let hostName = System.Net.Dns.GetHostName()